Simplify API without magnitude and kwargs (pt1)
As this commit might be scary and confusing, here are some explanations.

After discussing with @agramfor, we agreed that the Transform API up to
now had some unecessary complexity that was useful in some of the teams
works, but which might be confusing for braindecode users and
contributors. These include mainly the magnitude and mag_range
parameters, as well as the kwargs. Besides the clarity and
over-complexity issue, the kwargs posed an additional problem. While it
is useful to enable the generic Transform class to handle arbitrary
`operation` functions, which may require different types
of arguments, it is difficult to document and potentially
dangerous (as some kwargs may be mispelled and unused within the
functionals).

In order to prevent both these problems, we propose in this commit to
drop unecessary arguments and to get inspiration from the `torchvision`
library to handle operations with different signatures. More precisely:
1) we drop the `magnitude` and `mag_range` parameters;
2) we remove most of the random samplings from operations defined within
`functional.py`, leaving it to the classes defined within `transforms.py`;
3) we use a new get_params method within each class
inheriting from `Transform` to handle operations with different
signatures (as well as the random samplings).

As this change will require some work, I am first commiting this a
version where it was implemented for two augmentations (FrequencyShift
and TimeReverse) so that we can discuss and converge into something
before we implement it for the other augmentations.

Let me know what you think :)