Fix device handling in augmentations and tests
In commit 3f18ee9d63cd081b969f8a7e0ab1f3b9019000a3, functionals defining
Transforms are slightly modified to avoid non differentiable operations
wrt some of its hyperparameters (interesting for automatic data
augmentation for example). The problem is that this would break the
tests when running them in a CUDA environmnent (which was not noticed).
This commit fixes this (the corresponding tests in CUDA were already
commited).