Fix torch.linalg issue in rotation augmentations (#398)
* Fix torch.linalg issue in rotation augmentations

This commit fixes an issue encountered in the implementation of
augmentations using the sensors_rotation functional. The latter comes
from a bug that appeared in torch.linalg.pinv function since its
reimplementation for v1.8.0 release. The v1.8.0 implementation indeed
led too often in the failing of computation of the pseudo-inverse,
leading to incorrect augmentations at times and irreproducible runs.

The fix I propose in this commit follows recomendations in this torch issue
https://github.com/pytorch/pytorch/issues/75494 and consists in:
1. trying to use regular inverse whenever possible (since the matrix C
   in _torch_make_interpolation_matrix is supposed to be invertible
   quite often).
2. falling back to pinv if C is not invertible, but doing the
   computation in cpu (which is more reliable according to the issue
   above).

* Remove print

* Add deleted test back

* Revert "Add deleted test back"

This reverts commit a12742cfa7edf7dc4fe81ea701e75effa02fdddd.